completed studio for orm1
