+++ spring/codingevents/src/main/java/org/launchcode/codingevents/controllers/EventCategoryController.java
@@ -1,6 +1,7 @@
 package org.launchcode.codingevents.controllers;
 
 
+import jakarta.validation.Valid;
 import org.launchcode.codingevents.data.EventCategoryRepository;
 import org.launchcode.codingevents.models.Event;
 import org.launchcode.codingevents.models.EventCategory;
@@ -23,8 +24,8 @@ public class EventCategoryController {
 
     @GetMapping()
     public String displayAllCategories(Model model) {
-        model.addAttribute("categories", eventCategoryRepository.findAll());
         model.addAttribute("title", "All Categories");
+        model.addAttribute("categories", eventCategoryRepository.findAll());
 
         return "eventCategories/index";
     }
@@ -38,7 +39,7 @@ public class EventCategoryController {
     }
 
     @PostMapping("create")
-    public String processCreateEventCategoryForm(@ModelAttribute EventCategory newEventCategory, Errors errors, Model model) {
+    public String processCreateEventCategoryForm(@Valid @ModelAttribute EventCategory newEventCategory, Errors errors, Model model) {
         if(errors.hasErrors()){
             model.addAttribute("title", "Create Category");
             model.addAttribute("category", new EventCategory());
@@ -46,7 +47,7 @@ public class EventCategoryController {
         }
 
         eventCategoryRepository.save(newEventCategory);
-        return "eventCategories/create";
+        return "redirect:/eventCategories";
 
     }
 }
