+++ spring/codingevents/src/main/java/org/launchcode/codingevents/models/EventCategory.java
@@ -3,16 +3,17 @@ package org.launchcode.codingevents.models;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
+import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
 
 import java.util.Objects;
 
 @Entity
-public class EventCategory {
-    @Id
-    @GeneratedValue
-    private int id;
+public class EventCategory extends AbstractEntity{
+//    @Id
+//    @GeneratedValue
+//    private int id; gone to abstractentity
 
     @Size(min=3, message="Name must be at least 3 characters long")
     @NotBlank(message = "Input is required.")
@@ -23,13 +24,10 @@ public class EventCategory {
     }
     public EventCategory(){}
 
-    public int getId() {
-        return id;
-    }
+//    public int getId() {
+//        return id;
+//    } gone to abstractentity
 
-    public void setId(int id) {
-        this.id = id;
-    }
 
     public String getName() {
         return name;
@@ -39,18 +37,18 @@ public class EventCategory {
         this.name = name;
     }
 
-    @Override
-    public int hashCode() {
-        return Objects.hash(id);
-    }
-
-    @Override
-    public boolean equals(Object o) {
-            if (this == o) return true;
-            if (o == null || getClass() != o.getClass()) return false;
-            EventCategory that = (EventCategory) o;
-            return id == that.id;
-        }
+//    @Override
+////    public int hashCode() {
+////        return Objects.hash(id);
+////    }
+////
+////    @Override
+////    public boolean equals(Object o) {
+////            if (this == o) return true;
+////            if (o == null || getClass() != o.getClass()) return false;
+////            EventCategory that = (EventCategory) o;
+////            return id == that.id;
+//        } gone to abstractentity
 
 
     @Override
