+++ spring/codingevents/src/main/resources/templates/events/delete.html
@@ -4,6 +4,7 @@
 
 <body class="container">
     <title>Delete Event</title>
+    <nav th:replace = "~{fragments :: navigation}"></nav>
 
     <form method="post">
